Fixed WpiLib I2C File
Dearest Garry, There is a bug in WPILIB I2C file, so I created a
modified version in our project. The bug has to due with the read()
function saying it was aborted when it was not. Please work on getting
the driving function working better. I will be back at 7:45